requirements.txt
dataframe_image
docopt
imageio
imutils
kornia
kornia_rs
matplotlib
ml_metrics
numpy
opencv_contrib_python
opencv_python
opencv_python_headless
pandas
Pillow
python_Levenshtein
scikit_learn
scipy
toolz
tqdm
